move voice picking to its own file
